Add nChainWork to snapshot
This is a placeholder for the chain_stake.
Reasoning is described here: https://github.com/dtr-org/unit-e/issues/436

While working on this PR noticed these 2 issues:
1. There are multiple messages that have the same or partially the same schema.
Extracting the common fields into a separate struct would make it easier to
extend the schema of the snapshot overall and make sure that all relevant parts are updated.
2. Most of the functions accept raw fields, e.g.:
```c++
Indexer::Indexer(const uint256 &snapshot_hash, const uint256 &block_hash,
                 const uint256 &stake_modifier, const uint256 &chain_work,
                 uint32_t step, uint32_t steps_per_file)
```
It's easy to pass arguments in the wrong order and introduce a bug.
Would be better to pass objects that already hold these values,
e.g. `CBlockIndex` instead of `block_hash`, `stake_modifier` and `chain_work`.

I intend to work on the refactoring and tackling the above issues in the follow-up PR.

+++ src/test/snapshot/iterator_tests.cpp
@@ -19,7 +19,7 @@ BOOST_AUTO_TEST_CASE(snapshot_iterator) {
   {
     // generate the snapshot
     uint256 blockHash = uint256S("aa");
-    snapshot::Indexer idx(uint256S("bb"), blockHash, uint256(), 3, 2);
+    snapshot::Indexer idx(uint256S("bb"), blockHash, uint256(), uint256(), 3, 2);
     for (uint32_t i = 0; i < msgsToGenerate; ++i) {
       snapshot::UTXOSubset subset;
       subset.tx_id.SetHex(std::to_string(i));
@@ -39,8 +39,8 @@ BOOST_AUTO_TEST_CASE(snapshot_iterator) {
     std::unique_ptr<snapshot::Indexer> idx = snapshot::Indexer::Open(uint256S("bb"));
     BOOST_CHECK(idx != nullptr);
     snapshot::Iterator iter(std::move(idx));
-    BOOST_CHECK_EQUAL(iter.CalculateHash(uint256()).GetHex(),
-                      snapshotHash.GetHash(uint256()).GetHex());
+    BOOST_CHECK_EQUAL(iter.CalculateHash(uint256(), uint256()).GetHex(),
+                      snapshotHash.GetHash(uint256(), uint256()).GetHex());
   }
 
   {
